Add files via upload
I forgot to add a meaningful message to my last upload so I'm adding it to this one. This is a calculator to go along with the previous calculator, this one is able to take in a three arguments to determine how much to tip a meal and how to split the bill, while the last one calculates a 2d array of integers and adds one to any integers within that array via arguments, once and invalid argument is past, it terminates and prints the array.